% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes-data.R
\docType{data}
\name{diabetes_data}
\alias{diabetes_data}
\title{Diabetes}
\format{\describe{
  \item{sampleID}{??}
  \item{case}{??}
  \item{time}{??}
}}
\source{
\url{??}
}
\usage{
diabetes_data
}
\description{
A data frame with 768 rows and 9 variables
}
\keyword{datasets}
