% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBM_boot_err_Func.R
\name{GBM.boot.err.Func}
\alias{GBM.boot.err.Func}
\title{Application of GBM method}
\usage{
GBM.boot.err.Func(train, test, depVar, distribution = "adaboost",
  n.trees = 500)
}
\arguments{
\item{kernel}{A kernel function to be used.}

\item{cost}{A cost?}
}
\value{
A GBM.err a matrix of the estimated accurasy measures of Sensitivity, Specificity, and Misclassification for Training set, Testing set and overall one.
}
\description{
Returns the estimated accurasy measures of Sensitivity, Specificity, and Misclassification using Confusion matrix.
}
\details{
A GBM model is trained using a training data set and estimate accuracy measures using a test data set. The number of tree is defined by a user and the default value of it is 300.
}
