% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marine_pheno.R
\docType{data}
\name{Marine_pheno}
\alias{Marine_pheno}
\title{Marine_pheno}
\format{\describe{
  \item{sampleID}{??}
  \item{case}{??}
  \item{time}{??}
}}
\source{
\url{??}
}
\usage{
Marine_pheno
}
\description{
A data frame with 188 rows and 3 variables
}
\keyword{datasets}
