% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marine_raw.R
\docType{data}
\name{Marine_raw}
\alias{Marine_raw}
\title{Marine_raw}
\format{\describe{
  \item{Columns}{188 variables starting with GSM}
}}
\source{
\url{??}
}
\usage{
Marine_raw
}
\description{
A data frame with 27,974 rows and 188 variables
}
\keyword{datasets}
