% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_error_function.R
\name{ML_error}
\alias{ML_error}
\title{Application of each ML method to a testing set}
\usage{
ML_error(data, newdata, importance_measure, add.vars = NULL, rdit = 4,
  depVar = depVar, indVar = indVar, ML = c("RF", "BOOST",
  "Logistic"))
}
\arguments{
\item{data}{A matrix of the whole training data set for coefficients' esimates}

\item{newdata}{A matrix of a testing data}

\item{importance_measure}{A vector of importance measures obtained from Random Forest model}

\item{ML}{A ML method to apply}

\item{indvar}{A list of candidate predictors}

\item{depvar}{An outcome variable}
}
\value{
ML.err.vect A data frame including estimated accuracy measures
}
\description{
Returns the estimated accurasy measures of Sensitivity, Specificity, and Misclassification using Confusion matrix.
}
\details{
This is a generic function. The candidate predictors are a series of forward-in variables which can be picked by the variables ranked by
 importance measures or user-defined ones in `depvar` parameter.
 Their coefficients are estimated or the decision tree is built by by the whole training data.
}
