% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NN_boot_err_Func.R
\name{NN.boot.err.Func}
\alias{NN.boot.err.Func}
\title{Application of Neural Network method}
\usage{
NN.boot.err.Func(train, test, depVar, hidden = 3,
  linear.output = FALSE)
}
\arguments{
\item{train}{??}

\item{test}{??}

\item{depVar}{??}

\item{hidden}{A numeric value of # of hidden layers}

\item{linear.output}{??}
}
\value{
A NN.err a matrix of the estimated accurasy measures of Sensitivity, Specificity, and Misclassification for Training set, Testing set and overall one.
}
\description{
Returns the estimated accurasy measures of Sensitivity, Specificity, and Misclassification using Confusion matrix.
}
\details{
A Neural Network model is trained using a training data set and estimate accuracy measures using a test data set. The number of tree is defined by a user and the default value of it is 300.
}
