% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PR_boot_err_Func.R
\name{PR.boot.err.Func}
\alias{PR.boot.err.Func}
\title{Application of Penalized Regression method}
\usage{
PR.boot.err.Func(x.train, y.train, x.test, y.test, alphaLevel,
  family = "binomial", type = "class")
}
\arguments{
\item{x.train}{A data frame of training data set for independent variables.}

\item{y.train}{A data frame of training data set for dependent variable.}

\item{x.test}{A data frame of testing data set for independent variables.}

\item{y.test}{A data frame of testing data set for dependent variable.}

\item{alphaLevel}{User defined Alpha tunning parameter setting.}

\item{family}{A distribution of the outcome variable.}

\item{type}{Is this a classification or regression?}
}
\value{
A list of PR.fit PR fitted object. A non0coeff is the estimated non-zerio coefficients. A PR.err.vect a matrix of the estimated accurasy measures of Sensitivity, Specificity, and Misclassification for Training set, Testing set and overall one.
}
\description{
Returns the estimated accurasy measures of Sensitivity, Specificity, and Misclassification using Confusion matrix.
}
\details{
A Penalized Regression model is trained using a training data set and estimate accuracy measures using a test data set. The number of tree is defined by a user and the default value of it is 300.
}
