% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_boot_err_Func.R
\name{RF.boot.err.Func}
\alias{RF.boot.err.Func}
\title{Application of Random Forest method}
\usage{
RF.boot.err.Func(train, test, depVar, ntree = 300)
}
\arguments{
\item{train}{A data frame of training data set}

\item{test}{A data frame of a testing data}

\item{depVar}{An outcome variable}

\item{ntree}{A numeric value of # of trees to build}
}
\value{
A list of RF.fit RF fitted object, RF.err a matrix of the estimated accurasy measures of Sensitivity, Specificity, Misclassification for Training set, Testing set and overall one,
  importance_measure as a matrix form of Mean.Decrease.Accuracy and Mean.Decrease.Gini
}
\description{
Returns the estimated accurasy measures of Sensitivity, Specificity, and Misclassification using Confusion matrix.
}
\details{
A random forest model is trained using a training data set and estimate accuracy measures using a test data set. The number of tree is defined by a user and the default value of it is 300.
}
