% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVM_boot_err_Func.R
\name{SVM.boot.err.Func}
\alias{SVM.boot.err.Func}
\title{Application of SVM method}
\usage{
SVM.boot.err.Func(train, test, depVar, kernel = "radial", cost = 5)
}
\arguments{
\item{train}{??}

\item{test}{??}

\item{depVar}{??}

\item{kernel}{A kernel function to be used.}

\item{cost}{A cost?}
}
\value{
A SVM.err a matrix of the estimated accurasy measures of Sensitivity, Specificity, and Misclassification for Training set, Testing set and overall one.
}
\description{
Returns the estimated accurasy measures of Sensitivity, Specificity, and Misclassification using Confusion matrix.
}
\details{
A SVM model is trained using a training data set and estimate accuracy measures using a test data set. The number of tree is defined by a user and the default value of it is 300.
}
