% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WTC_pheno.R
\docType{data}
\name{WTC_pheno}
\alias{WTC_pheno}
\title{WTC_pheno}
\format{\describe{
  \item{STUDYID}{??}
  \item{PTSD}{??}
}}
\source{
\url{??}
}
\usage{
WTC_pheno
}
\description{
A data frame with 324 rows and 2 variables
}
\keyword{datasets}
