% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WTC_raw.R
\docType{data}
\name{WTC_raw}
\alias{WTC_raw}
\title{WTC_raw}
\format{\describe{
  \item{Columns}{324 variables starting with id_}
}}
\source{
\url{??}
}
\usage{
WTC_raw
}
\description{
A data frame with 25,830 rows and 324 variables
}
\keyword{datasets}
